fix(ai/llm): repair 3 failing batch5 happy-path tests (BIT-371)
Three failures in ai_llm_workflow_batch5_tests.rs returned 500 instead of 200:

- get_ai_statistics: get_usage_statistics typed COALESCE(AVG(latency_ms), 0)
  as f64, but Postgres AVG(integer) yields numeric, which sqlx cannot decode
  into f64 -> always 500. Added the ::float8 cast used by every other stats
  repo. Real handler bug (the totals row always exists), not test-only.
- get_lease_template: seed_prompt_template left provider/model NULL, but the
  LlmPromptTemplate struct types them as non-null String -> row decode failed.
  Seed now sets provider/model. (list test passed only because it returns 0 rows.)
- update_escalation_config: PUT handler runs a plain UPDATE ... RETURNING *
  (no upsert); with no pre-existing row fetch_one hit RowNotFound. Seed an
  ai_escalation_configs row to establish the precondition (prod auto-creates on GET).

+++ backend/crates/db/src/repositories/llm_document.rs
@@ -1766,7 +1766,7 @@ impl LlmDocumentRepository {
                 COUNT(*) FILTER (WHERE status = 'failed'),
                 COALESCE(SUM(tokens_used), 0),
                 COALESCE(SUM(cost_cents), 0),
-                COALESCE(AVG(latency_ms), 0)
+                COALESCE(AVG(latency_ms), 0)::float8
             FROM llm_generation_requests
             WHERE organization_id = $1
               AND created_at >= $2
@@ -1808,7 +1808,7 @@ impl LlmDocumentRepository {
                 COUNT(*),
                 COALESCE(SUM(tokens_used), 0),
                 COALESCE(SUM(cost_cents), 0),
-                COALESCE(AVG(latency_ms), 0)
+                COALESCE(AVG(latency_ms), 0)::float8
             FROM llm_generation_requests
             WHERE organization_id = $1
               AND created_at >= $2
